requirements.txt
appdirs==1.4.3
arrow==0.12.1
asn1crypto==0.24.0
attrs==18.2.0
blinker==1.4
cached-property==1.5.1
Cerberus==1.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
defusedxml==0.5.0
Eve==0.8.2.dev0
Eve-Swagger==0.0.11
Events==0.3
Flask==1.0.2
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.0
MarkupSafe==1.1.0
Pillow==5.4.0
pycparser==2.19
pygraphviz==1.5
PyJWT==1.7.1
pymongo==3.7.2
pytaf==1.2.1
python-dateutil==2.7.5
python-metar==1.4.0
python-yr==1.4.5
pytz==2018.7
requests==2.21.0
requests-toolbelt==0.8.0
simplejson==3.16.0
six==1.12.0
transitions==0.6.9
urllib3==1.24.1
Werkzeug==0.14.1
xmltodict==0.11.0
zeep==3.2.0
